+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+- Reduce number of render item calls
+  - https://github.com/Shopify/flash-list/pull/253
+
 ## [0.4.1] - 2022-03-29
 
 - Crash fix for android activity switching (#256)